+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Concept documentation for how Puppet invokes Puppetized DSC Resources in the `about_Puppetized_module_invocation` about help topic ([#74](https://github.com/puppetlabs/Puppet.Dsc/issues/74))
+- `Export-PuppetModule` function to build a Puppetized PowerShell module into a packaged Puppet module ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
+- `Get-ForgeModuleInfo` function for retrieving releases and PowerShell metadata about Puppetized PowerShell modules on the forge ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
+- `Get-PowerShellDscModule` function for discovering Puppetizable PowerShell modules in a PowerShell repository ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
+- `Update-ForgeDscModule` function for rebuilding and publishing (with incremented Puppet build version) Puppetized DSC modules to the Forge ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
+- `Publish-NewDscModuleVersion` function for Puppetizing and publishing versions of PowerShell modules with DSC Resources which are not yet on the Forge in a specified namespace ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
+
+### Changed
+
+- `Publish-PuppetModule` function to include the new `ExportFolder`, `PackagedModulePath`, and `Force` parameters, as well as refactor the function to use `Export-PuppetModule` and clean up the experience simultaneously building and publishing a Puppet module. Finally, the `ForgeToken` parameter now defaults to the `FORGE_TOKEN` environment variable if it is not passed directly ([#68](https://github.com/puppetlabs/ruby-pwsh/issues/68))
 
 ### Fixed